refactor: extract festival job header hook to satisfy size budget
The review fixes pushed FestivalArtistManagement.tsx past its file-size
ceiling (812 -> 826), failing the governance gate. Move the job header
loading (title, festival dates with route-date preselection, max
stages, offline race) into useFestivalArtistJobDetails; the page drops
under the 800-line threshold and the behavior is unchanged.
